# openapi.yml
openapi: 3.0.1
info:
  title: DIA Engine
  description: Decentralized Intelligence Agency
  termsOfService: http://swagger.io/terms/
  contact:
    email: help.bitskyai@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: http://localhost:9099/
...
paths:
...
  /apis/producers:
    get:
      tags:
        - producer
      description: Operation Index - 0012
      operationId: getProducers
      responses:
        200:
          description: Get Producers successful
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "./src/schemas/producer.json"
    post:
      tags:
        - producer
      summary: Register an Producer.
      description: Operation Index - 0013
      operationId: registerProducer
      requestBody:
        description: Need to register Producer information
        required: true
        content:
          application/json:
            schema:
              $ref: "./src/schemas/producer.json"
      responses:
        200:
          description: Registered successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id:
                    type: string
                    description: Generate MongoDB Id
                  global_id:
                    $ref: "#/components/schemas/GlobalId"
        400:
          description: Bad request, the data pass isn't valid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /apis/producers/{gid}:
    get:
      tags:
        - producer
      summary: Get an Producer
      description: Operation Index - 0014
      operationId: getProducer
      parameters:
        - in: path
          name: gid
          required: true
          description: global_id of Producer
          schema:
            $ref: "#/components/schemas/GlobalId"
      responses:
        200:
          description: Get an producer successful
          content:
            application/json:
              schema:
                $ref: "./src/schemas/producer.json"
        400:
          description: Bad request, the data pass isn't valid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        404:
          description: Cannot find specific Producer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      tags:
        - producer
      summary: Update an Producer
      description: Operation Index - 0015
      operationId: updateProducer
      parameters:
        - in: path
          name: gid
          required: true
          description: global_id of Producer
          schema:
            $ref: "#/components/schemas/GlobalId"
      requestBody:
        description: Producer information need to update
        required: true
        content:
          application/json:
            schema:
              $ref: "./src/schemas/producer.json"
      responses:
        204:
          description: Update Producer successful
        404:
          description: Cannot find specific Producer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags:
        - producer
      summary: Unregister an Producer
      description: Operation Index - 0016
      operationId: unregisterProducer
      parameters:
        - in: path
          name: gid
          required: true
          description: global_id of Producer
          schema:
            $ref: "#/components/schemas/GlobalId"
      responses:
        200:
          description: Delete Producer successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  global_id:
                    $ref: "#/components/schemas/GlobalId"
        404:
          description: Cannot find specific Producer
          content:
            application/json:
              schema:
                type: object
                properties:
                  global_id:
                    $ref: "#/components/schemas/GlobalId"
...
components:
  schemas:
    GlobalId:
      type: string
      description: Global unique id.
      example: ef72448b-7bc4-42d1-9436-822c9634d943
...
    Error:
      type: object
      properties:
        stack:
          type: string
          description: Error stack, return to UI only in develop mode
        message:
          type: string
          description: Error Message
        statusCode:
          type: number
          description: HTTP Status Code
        code:
          type: string
          description: Error Code
        data:
          type: object
          description: Additional data for this error
  securitySchemes:
    api_key:
      type: apiKey
      name: X-API-KEY
      in: header